// Products.WebApp/src/app/api/client.ts
export class CreateProductCommand implements ICreateProductCommand {
    name?: string | undefined;
    color?: string | undefined;
    price?: number;

    constructor(data?: ICreateProductCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.color = _data["color"];
            this.price = _data["price"];
        }
    }

    static fromJS(data: any): CreateProductCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateProductCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["color"] = this.color;
        data["price"] = this.price;
        return data;
    }
}

export interface ICreateProductCommand {
    name?: string | undefined;
    color?: string | undefined;
    price?: number;
}

export class GetProductsQuery implements IGetProductsQuery {
    filter?: { [key: string]: string; } | undefined;
    sort?: string | undefined;
    ascending?: boolean;
    page?: number;
    pageSize?: number;

    constructor(data?: IGetProductsQuery) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (_data["filter"]) {
                this.filter = {} as any;
                for (let key in _data["filter"]) {
                    if (_data["filter"].hasOwnProperty(key))
                        (this.filter as any)![key] = _data["filter"][key];
                }
            }
            this.sort = _data["sort"];
            this.ascending = _data["ascending"];
            this.page = _data["page"];
            this.pageSize = _data["pageSize"];
        }
    }

    static fromJS(data: any): GetProductsQuery {
        data = typeof data === 'object' ? data : {};
        let result = new GetProductsQuery();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (this.filter) {
            data["filter"] = {};
            for (let key in this.filter) {
                if (this.filter.hasOwnProperty(key))
                    (data["filter"] as any)[key] = (this.filter as any)[key];
            }
        }
        data["sort"] = this.sort;
        data["ascending"] = this.ascending;
        data["page"] = this.page;
        data["pageSize"] = this.pageSize;
        return data;
    }
}

export interface IGetProductsQuery {
    filter?: { [key: string]: string; } | undefined;
    sort?: string | undefined;
    ascending?: boolean;
    page?: number;
    pageSize?: number;
}
// ...
export class ProductDto implements IProductDto {
    id?: number;
    name!: string | undefined;
    color!: string | undefined;
    price?: number;

    constructor(data?: IProductDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.color = _data["color"];
            this.price = _data["price"];
        }
    }

    static fromJS(data: any): ProductDto {
        data = typeof data === 'object' ? data : {};
        let result = new ProductDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["color"] = this.color;
        data["price"] = this.price;
        return data;
    }
}

export interface IProductDto {
    id?: number;
    name: string | undefined;
    color: string | undefined;
    price?: number;
}

export class ProductDtoListGridResult implements IProductDtoListGridResult {
    data?: ProductDto[][] | undefined;
    total?: number;
    page?: number;

    constructor(data?: IProductDtoListGridResult) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["data"])) {
                this.data = [] as any;
                for (let item of _data["data"])
                    this.data!.push(item);
            }
            this.total = _data["total"];
            this.page = _data["page"];
        }
    }

    static fromJS(data: any): ProductDtoListGridResult {
        data = typeof data === 'object' ? data : {};
        let result = new ProductDtoListGridResult();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.data)) {
            data["data"] = [];
            for (let item of this.data)
                data["data"].push(item);
        }
        data["total"] = this.total;
        data["page"] = this.page;
        return data;
    }
}

export interface IProductDtoListGridResult {
    data?: ProductDto[][] | undefined;
    total?: number;
    page?: number;
}

export class ProductDtoListGridResultApiResponse implements IProductDtoListGridResultApiResponse {
    success?: boolean;
    message?: string | undefined;
    data?: ProductDtoListGridResult;
    errors?: string[] | undefined;

    constructor(data?: IProductDtoListGridResultApiResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.success = _data["success"];
            this.message = _data["message"];
            this.data = _data["data"] ? ProductDtoListGridResult.fromJS(_data["data"]) : undefined as any;
            if (Array.isArray(_data["errors"])) {
                this.errors = [] as any;
                for (let item of _data["errors"])
                    this.errors!.push(item);
            }
        }
    }

    static fromJS(data: any): ProductDtoListGridResultApiResponse {
        data = typeof data === 'object' ? data : {};
        let result = new ProductDtoListGridResultApiResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["success"] = this.success;
        data["message"] = this.message;
        data["data"] = this.data ? this.data.toJSON() : undefined as any;
        if (Array.isArray(this.errors)) {
            data["errors"] = [];
            for (let item of this.errors)
                data["errors"].push(item);
        }
        return data;
    }
}

export interface IProductDtoListGridResultApiResponse {
    success?: boolean;
    message?: string | undefined;
    data?: ProductDtoListGridResult;
    errors?: string[] | undefined;
}
// ...
export class UpdateProductCommand implements IUpdateProductCommand {
    id?: number;
    name?: string | undefined;
    color?: string | undefined;
    price?: number;

    constructor(data?: IUpdateProductCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.color = _data["color"];
            this.price = _data["price"];
        }
    }

    static fromJS(data: any): UpdateProductCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateProductCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["color"] = this.color;
        data["price"] = this.price;
        return data;
    }
}

export interface IUpdateProductCommand {
    id?: number;
    name?: string | undefined;
    color?: string | undefined;
    price?: number;
}